+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/ConfigurationImpl.kt
@@ -21,6 +21,7 @@ import io.realm.kotlin.EncryptionKeyCallback
 import io.realm.kotlin.InitialDataCallback
 import io.realm.kotlin.InitialRealmFileConfiguration
 import io.realm.kotlin.LogConfiguration
+import io.realm.kotlin.annotations.ExperimentalEncryptionCallbackApi
 import io.realm.kotlin.dynamic.DynamicMutableRealm
 import io.realm.kotlin.dynamic.DynamicMutableRealmObject
 import io.realm.kotlin.dynamic.DynamicRealm
@@ -61,6 +62,7 @@ public open class ConfigurationImpl(
     schemaVersion: Long,
     schemaMode: SchemaMode,
     private val userEncryptionKey: ByteArray?,
+    @OptIn(ExperimentalEncryptionCallbackApi::class)
     override val encryptionKeyAsCallback: EncryptionKeyCallback?,
     compactOnLaunchCallback: CompactOnLaunchCallback?,
     userMigration: RealmMigration?,
@@ -232,6 +234,7 @@ public open class ConfigurationImpl(
                 RealmInterop.realm_config_set_encryption_key(nativeConfig, key)
             }
 
+            @OptIn(ExperimentalEncryptionCallbackApi::class)
             encryptionKeyAsCallback?.let {
                 RealmInterop.realm_config_set_encryption_key_from_pointer(nativeConfig, it.keyPointer())
             }